requirements.txt
pygame
textual
pillow
textual-image
